Voice/requirements.txt
# Web Scraping
scrapy==2.11.0
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.24.3
scipy==1.11.4

# NLP & Text Processing
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
nltk==3.8.1
gensim==4.3.2
textstat==0.7.3
vaderSentiment==3.3.2

# Audio Processing
librosa==0.10.1
pydub==0.25.1
speechrecognition==3.10.0
praat-parselmouth==0.4.3

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
tensorflow==2.15.0

# Visualization & Dashboard
streamlit==1.28.1
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Explainable AI
shap==0.44.0
lime==0.2.0.1

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
yfinance==0.2.28
pytz==2023.3
